✨  CellItem에서 사용되는 css함수 helper로 뺌
